Construct the fundamental layer of the API service
Architecture:

Since the data relationship has become complicated, I changed to use Relational Database Service and
closed DynamoDB. It means that most Lambdas will have to interact with RDS.

After some research on Cognito service, I found it does not match our needs as our target platform
is IOS instead of Web. Therefore, I abandoned Cognito and write an Lambda for API authorizer.

Besides the project name, stage, and the certificate ARN, one should give a top domain, an API
domain (usually a subdomain under the top domain), a database username, a database password, and a
Lambda Role ARN when deploying this template.

Notice that it's not a best practice to put a database in public network and to save sensitive
information in environment variables. However, it's OK for testing.

Move Lambdas into a VPC public subnet and put the database into a VPC private subnet can solve the
first problem.
For the second problem, Secret Manager will be the solution.

Lambda Layers:

Currently, there are two lambda layers are used: one is 'common' layer and the other is 'oauth'
layer.

For not packaging the monumental library code and increasing reusability, I put the packages that
are common used by most lambdas in 'common' lambda layer and put the Google OAuth package in 'oauth'
lambda layer.

To prevent Webpack from including the library when building the project, I adjusted the config file.

Before packaging the project for deployment, run the script 'npm run package-layer' to package the
layer files first.

Database initialization:

Because CloudFormation is incapable of initializing tables in the database, the process is needed to
be performed manually or programmatically.

Here, I wrote a file 'db-init.sql' for initializing all tables in a database.

It works by entering the command 'mysql -h DB_HOST -p 3306 -u DB_USERNAME -p DB_PASSWORD
ramen_pocket < db-init.sql'.

Again, it's not the best practice to initialize a database due to the security issue.

Authorizer Lambda:

The lambda 'authenticator' is used to verify the OAuth token in a request's headers and respond with
access policies to the API Gateway. Therefore the API Gateway will know which resources the user has
or not has permissions to access.