Extend class DatabaseConnection and its related types
Change the interface Options to the class DatabaseConnectionOptions for a clearer intention.

class Counter is defined for a type of commonly used query.

The return type of method query of DatabaseConnection now depends on the generics notation T.

The return type of method batch of DatabaseConnection is now an array of UpsertResult from package
mariadb.

Method beginTransaction, commit, rollback, runTransactionTask, and checkConnection are new to
DatabaseConnection.

Types errors in the files depending on DatabaseConnection is also fixed.